test(logs): fix quota stat scan overwrite and plugin e2e assertion type
- SumUsedQuotaWithModelNameMode scanned quota and rpm/tpm into the same
  Stat, letting the second scan zero out Quota under the upgraded GORM;
  aggregate quota into a separate stat and restore it after the rpm/tpm scan.
- plugin native e2e asserted User.Quota (int64) against an untyped int,
  failing testify's strict type check.

+++ controller/plugin_native_e2e_test.go
@@ -218,7 +218,7 @@ func TestKlingNativeRouteSubmitPollSettleAndQuery(t *testing.T) {
 	assert.Equal(t, int32(1), queryCalls.Load())
 	var settledUser model.User
 	require.NoError(t, database.First(&settledUser, 7).Error)
-	assert.Equal(t, 999_999, settledUser.Quota)
+	assert.Equal(t, int64(999_999), settledUser.Quota)
 
 	queryBinding, found := generation.LookupDeclaredRoute(http.MethodGet, "/kling/v1/videos/text2video/:task_id")
 	require.True(t, found)
